dags/requirements.txt
# Airflow requirements
apache-airflow-providers-cncf-kubernetes>=7.4.0
apache-airflow-providers-amazon>=8.13.0
boto3>=1.34.0
kubernetes>=28.1.0
